ExC v0.2.0 ExC

Main entry point that handles all the compiling process.

Functions

main(argv)

Filters command line arguments.

Specs

argv can take any of the following values:

-h or --help : displays help regarding the use of the compiler

-v or --verbose : executes the compiler with logs on each step

-v or --verbose : executes the compiler with logs on each step

Examples

  $ ./exc <file_to_compile>

start_compilation(file_path, verbose)

Starts the compilation process given an input path.

Specs

file_path is the path to the file to be compiled. verbose boolean to indicate if there should be a detailed output of each of the compiler's stages.

Examples

  $ ./exc examples/test.c